Start working on applications
