+++ www/src/components/shell/onboarding/OnboardingFlow.js
@@ -62,15 +62,15 @@ function OnboardingFlow({ accessToken, provider: scmProvider, authUrlData }) {
   }, [setProvider, setCredentials, setWorkspace, workspace])
 
   const next = useCallback(() => {
-    const hasNext = !!SECTIONS[section].next
-
-    if (hasNext) setSection(SECTIONS[section].next)
+    if (SECTIONS[section].next) {
+      setSection(SECTIONS[section].next)
+    }
   }, [section])
 
   const previous = useCallback(() => {
-    const hasPrevious = !!SECTIONS[section].previous
-
-    if (hasPrevious) setSection(SECTIONS[section].previous)
+    if (SECTIONS[section].previous) {
+      setSection(SECTIONS[section].previous)
+    }
   }, [section])
 
   const validations = getValidations(providerName, scmProvider, section)
