+++ www/src/components/shell/queries.js
@@ -1,6 +1,7 @@
 import { gql } from '@apollo/client'
 
 import { CloudShellFragment, DemoProjectFragment } from '../../models/shell'
+import { PageInfo } from '../../models/misc'
 
 export const AUTHENTICATION_URLS_QUERY = gql`
   query {
@@ -60,4 +61,21 @@ export const POLL_DEMO_PROJECT_QUERY = gql`
     }
   }
   ${DemoProjectFragment}
+`
+
+export const APPLICATIONS_QUERY = gql`
+  query ApplicationsQuery($cursor: String) {
+    repositories(after: $cursor, first: 200) {
+      pageInfo {
+        ...PageInfo
+      }
+      edges {
+        node {
+          id
+          name
+        }
+      }
+    }
+  }
+  ${PageInfo}
 `
